+++ src/interpreter/constructs.ts
@@ -1,5 +1,7 @@
+import { LocationSpan } from '@superfaceai/ast';
 import * as ts from 'typescript';
 
+import { IssueLocation } from '.';
 import { ValidationIssue } from './issue';
 import { ValidationResult } from './map-validator';
 import {
@@ -45,6 +47,7 @@ export type ConstructResult = ValidationResult & {
 export interface VisitConstruct<T extends ts.Node = ts.Node> {
   visit(
     node: T,
+    initialLocation: LocationSpan | undefined,
     outputStructure?: StructureType,
     inputStructure?: ObjectStructure,
     isOutcomeWithCondition?: boolean
@@ -88,8 +91,49 @@ function mergeResults(...results: ConstructResult[]): ConstructResult {
   }, VALID_CONSTRUCT_RESULT);
 }
 
-function getPath(node: ts.Node): string[] {
-  return [`${node.getStart()}:${node.getEnd()}`, ts.SyntaxKind[node.kind]];
+function getPath(
+  node: ts.Node,
+  initialLocation: LocationSpan | undefined
+): IssueLocation {
+  if (initialLocation) {
+    let lineOffset = 0;
+    let columnOffset = -2; // -2 because expression passed in parent source file is wrapped in parenthesis
+
+    const start = node
+      .getSourceFile()
+      .getLineAndCharacterOfPosition(node.getStart());
+
+    lineOffset = start.line + initialLocation.start.line;
+
+    // if jessie expression is on multiple lines - do not preserve initial column location
+    if (lineOffset > initialLocation.start.line) {
+      columnOffset += start.character;
+    } else {
+      columnOffset += start.character + initialLocation.start.column;
+    }
+
+    const end = node
+      .getSourceFile()
+      .getLineAndCharacterOfPosition(node.getEnd());
+
+    return {
+      kind: ts.SyntaxKind[node.kind],
+      location: {
+        start: {
+          line: start.line + lineOffset,
+          column: start.character + columnOffset,
+          charIndex: start.character - 1,
+        },
+        end: {
+          line: end.line + lineOffset,
+          column: end.character + columnOffset,
+          charIndex: end.character - 1,
+        },
+      },
+    };
+  } else {
+    return { kind: ts.SyntaxKind[node.kind] };
+  }
 }
 
 function isTypescriptIdentifier(node: ts.Node): node is TypescriptIdentifier {
@@ -103,6 +147,7 @@ function isTypescriptIdentifier(node: ts.Node): node is TypescriptIdentifier {
 
 function compareStructures(
   node: ts.Node,
+  initialLocation: LocationSpan | undefined,
   inputStructure: StructureType,
   outputStructure: StructureType
 ): ConstructResult {
@@ -163,7 +208,7 @@ function compareStructures(
       {
         kind: 'wrongStructure',
         context: {
-          path: getPath(node),
+          path: getPath(node, initialLocation),
           expected: outputStructure,
           actual: inputStructure,
         },
@@ -176,6 +221,7 @@ function compareStructures(
 
 function visitConstruct(
   node: ts.Node,
+  initialLocation: LocationSpan | undefined,
   outputStructure?: StructureType,
   inputStructure?: ObjectStructure,
   isOutcomeWithCondition?: boolean,
@@ -184,6 +230,7 @@ function visitConstruct(
   return construct
     ? construct.visit(
         node,
+        initialLocation,
         outputStructure,
         inputStructure,
         isOutcomeWithCondition
@@ -246,6 +293,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.StringLiteral]: {
     visit: (
       node: ts.StringLiteral,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       _inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -267,7 +315,7 @@ export const RETURN_CONSTRUCTS: {
         {
           kind: 'wrongStructure',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             actual: node.text,
             expected: outputStructure,
           },
@@ -282,6 +330,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.NumericLiteral]: {
     visit: (
       node: ts.NumericLiteral,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       _inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -303,7 +352,7 @@ export const RETURN_CONSTRUCTS: {
         {
           kind: 'wrongStructure',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             actual: node.text,
             expected: outputStructure,
           },
@@ -318,6 +367,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.FalseKeyword]: {
     visit: (
       node: ts.FalseLiteral,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       _inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -340,7 +390,7 @@ export const RETURN_CONSTRUCTS: {
         {
           kind: 'wrongStructure',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             actual: 'false',
             expected: outputStructure,
           },
@@ -355,6 +405,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.TrueKeyword]: {
     visit: (
       node: ts.TrueLiteral,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       _inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -377,7 +428,7 @@ export const RETURN_CONSTRUCTS: {
         {
           kind: 'wrongStructure',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             actual: 'true',
             expected: outputStructure,
           },
@@ -392,6 +443,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.NullKeyword]: {
     visit: (
       node: ts.NullLiteral,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       _inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -404,7 +456,7 @@ export const RETURN_CONSTRUCTS: {
           {
             kind: 'wrongStructure',
             context: {
-              path: getPath(node),
+              path: getPath(node, initialLocation),
               actual: 'null',
               expected: outputStructure,
             },
@@ -422,6 +474,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.BinaryExpression]: {
     visit: (
       node: ts.BinaryExpression,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -434,6 +487,7 @@ export const RETURN_CONSTRUCTS: {
           results.push(
             visitConstruct(
               node.left,
+              initialLocation,
               undefined,
               inputStructure,
               isOutcomeWithCondition,
@@ -445,6 +499,7 @@ export const RETURN_CONSTRUCTS: {
           results.push(
             visitConstruct(
               node.right,
+              initialLocation,
               undefined,
               inputStructure,
               isOutcomeWithCondition,
@@ -473,6 +528,7 @@ export const RETURN_CONSTRUCTS: {
       results.push(
         visitConstruct(
           node.left,
+          initialLocation,
           outputStructure,
           undefined,
           isOutcomeWithCondition,
@@ -480,6 +536,7 @@ export const RETURN_CONSTRUCTS: {
         ),
         visitConstruct(
           node.right,
+          initialLocation,
           outputStructure,
           undefined,
           isOutcomeWithCondition,
@@ -494,6 +551,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.Identifier]: {
     visit: (
       node: ts.Identifier,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -504,7 +562,7 @@ export const RETURN_CONSTRUCTS: {
             {
               kind: 'inputNotFound',
               context: {
-                path: getPath(node),
+                path: getPath(node, initialLocation),
                 actual: getVariableName(node),
               },
             },
@@ -515,7 +573,12 @@ export const RETURN_CONSTRUCTS: {
         }
 
         if (outputStructure) {
-          return compareStructures(node, inputStructure, outputStructure);
+          return compareStructures(
+            node,
+            initialLocation,
+            inputStructure,
+            outputStructure
+          );
         }
 
         return VALID_CONSTRUCT_RESULT;
@@ -529,7 +592,7 @@ export const RETURN_CONSTRUCTS: {
               {
                 kind: 'wrongStructure',
                 context: {
-                  path: getPath(node),
+                  path: getPath(node, initialLocation),
                   actual: getVariableName(node),
                   expected: outputStructure.value,
                 },
@@ -559,6 +622,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.PropertyAccessExpression]: {
     visit(
       node: ts.PropertyAccessExpression,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -573,7 +637,7 @@ export const RETURN_CONSTRUCTS: {
             {
               kind: 'inputNotFound',
               context: {
-                path: getPath(node),
+                path: getPath(node, initialLocation),
                 actual: getVariableName(node),
               },
             },
@@ -586,7 +650,7 @@ export const RETURN_CONSTRUCTS: {
         const issue: ValidationIssue = {
           kind: 'wrongInput',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             expected: inputStructure,
             actual: getVariableName(node),
           },
@@ -604,7 +668,12 @@ export const RETURN_CONSTRUCTS: {
         }
 
         if (outputStructure) {
-          return compareStructures(node, fieldValue, outputStructure);
+          return compareStructures(
+            node,
+            initialLocation,
+            fieldValue,
+            outputStructure
+          );
         }
 
         return VALID_CONSTRUCT_RESULT;
@@ -628,6 +697,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.ElementAccessExpression]: {
     visit(
       node: ts.ElementAccessExpression,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -641,7 +711,7 @@ export const RETURN_CONSTRUCTS: {
           const issue: ValidationIssue = {
             kind: 'inputNotFound',
             context: {
-              path: getPath(node),
+              path: getPath(node, initialLocation),
               actual: getVariableName(node),
             },
           };
@@ -652,7 +722,7 @@ export const RETURN_CONSTRUCTS: {
         const issue: ValidationIssue = {
           kind: 'wrongInput',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             expected: inputStructure,
             actual: getVariableName(node),
           },
@@ -670,7 +740,12 @@ export const RETURN_CONSTRUCTS: {
         }
 
         if (outputStructure) {
-          return compareStructures(node, fieldValue, outputStructure);
+          return compareStructures(
+            node,
+            initialLocation,
+            fieldValue,
+            outputStructure
+          );
         }
 
         return VALID_CONSTRUCT_RESULT;
@@ -694,6 +769,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.ObjectLiteralExpression]: {
     visit(
       node: ts.ObjectLiteralExpression,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -709,6 +785,7 @@ export const RETURN_CONSTRUCTS: {
             results.push(
               visitConstruct(
                 property.initializer,
+                initialLocation,
                 undefined,
                 inputStructure,
                 isOutcomeWithCondition,
@@ -735,7 +812,7 @@ export const RETURN_CONSTRUCTS: {
         const issue: ValidationIssue = {
           kind: 'wrongStructure',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             actual: node.getText(),
             expected: outputStructure,
           },
@@ -785,6 +862,7 @@ export const RETURN_CONSTRUCTS: {
         results.push(
           visitConstruct(
             property.initializer,
+            initialLocation,
             structureOfProperties[(property.name as ts.Identifier).text],
             undefined,
             isOutcomeWithCondition,
@@ -797,7 +875,7 @@ export const RETURN_CONSTRUCTS: {
         const issue: ValidationIssue = {
           kind: 'missingRequired',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             field: key,
           },
         };
@@ -809,7 +887,7 @@ export const RETURN_CONSTRUCTS: {
         const issue: ValidationIssue = {
           kind: 'wrongObjectStructure',
           context: {
-            path: getPath(node),
+            path: getPath(node, initialLocation),
             expected: structureOfProperties,
             actual: node.getText(),
           },
@@ -825,6 +903,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.ArrayLiteralExpression]: {
     visit(
       node: ts.ArrayLiteralExpression,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -837,6 +916,7 @@ export const RETURN_CONSTRUCTS: {
             results.push(
               visitConstruct(
                 element,
+                initialLocation,
                 undefined,
                 inputStructure,
                 isOutcomeWithCondition,
@@ -861,7 +941,7 @@ export const RETURN_CONSTRUCTS: {
       const wrongStructureIssue: ValidationIssue = {
         kind: 'wrongStructure',
         context: {
-          path: getPath(node),
+          path: getPath(node, initialLocation),
           actual: node.getText(),
           expected: outputStructure,
         },
@@ -888,6 +968,7 @@ export const RETURN_CONSTRUCTS: {
           results.push(
             visitConstruct(
               element,
+              initialLocation,
               structureOfType,
               undefined,
               isOutcomeWithCondition,
@@ -910,6 +991,7 @@ export const RETURN_CONSTRUCTS: {
           results.push(
             visitConstruct(
               element,
+              initialLocation,
               outputStructure.value,
               undefined,
               isOutcomeWithCondition,
@@ -924,6 +1006,7 @@ export const RETURN_CONSTRUCTS: {
         for (const value of typeValues) {
           const result = visitConstruct(
             element,
+            initialLocation,
             value,
             undefined,
             isOutcomeWithCondition,
@@ -954,12 +1037,14 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.ParenthesizedExpression]: {
     visit(
       node: ts.ParenthesizedExpression,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
     ): ConstructResult {
       return visitConstruct(
         node.expression,
+        initialLocation,
         outputStructure,
         inputStructure,
         isOutcomeWithCondition,
@@ -971,12 +1056,14 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.ExpressionStatement]: {
     visit(
       node: ts.ExpressionStatement,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
     ): ConstructResult {
       return visitConstruct(
         node.expression,
+        initialLocation,
         outputStructure,
         inputStructure,
         isOutcomeWithCondition,
@@ -988,6 +1075,7 @@ export const RETURN_CONSTRUCTS: {
   [ts.SyntaxKind.SourceFile]: {
     visit(
       node: ts.SourceFile,
+      initialLocation: LocationSpan | undefined,
       outputStructure?: StructureType,
       inputStructure?: ObjectStructure,
       isOutcomeWithCondition?: boolean
@@ -996,6 +1084,7 @@ export const RETURN_CONSTRUCTS: {
 
       return visitConstruct(
         statement,
+        initialLocation,
         outputStructure,
         inputStructure,
         isOutcomeWithCondition,